configure.ac use only pkg-config for proj library lookup
The current macro wood for libproj lookup is really complicated and
is known to be broken (eg. see bug#1871). It doesn't serve any really
practical purpose anyways, except for really ancient versions of proj
that might miss the .pc file (which probably won't work anyways).
We're already using pkg-config by default.

pkg-config is the well-proven and simple standard way for library
package lookups since over a decade, so that's all we need.

So just drop a hundred lines of unncessary code and canonicalize the
corresponding makefile variables.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>